pyproject.toml
[project]
name = "vector-store-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
]
